backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
numpy==1.26.4
pillow==10.3.0
tensorflow-cpu==2.16.1
opencv-python-headless==4.10.0.84
